Split into sessions and resources
