+++ app/education/models/details.py
@@ -1,46 +1,12 @@
-import wagtail.blocks
-from app.core.blocks.image import (
-    APIImageChooserBlock,
-    ImageGalleryBlock,
-    PartnerLogoChooserBlock,
-)
-from app.core.blocks.page_chooser import APIPageChooserBlock
-from app.core.blocks.paragraph import APIRichTextBlock
-from app.core.blocks.promoted_links import FeaturedExternalLinkBlock, FeaturedPageBlock
-from app.core.blocks.quote import QuoteBlock
-from app.core.blocks.section import SubHeadingBlock
-from app.core.blocks.video import YouTubeBlock
-from app.core.models import (
-    BasePageWithRequiredIntro,
-    PublishedDateMixin,
-    RequiredHeroImageMixin,
-)
-from app.core.serializers import RichTextSerializer
-from django.conf import settings
 from django.db import models
 from django.utils.functional import cached_property
 from django.utils.translation import gettext_lazy as _
-from modelcluster.fields import ParentalKey
-from wagtail.admin.panels import (
-    FieldPanel,
-    InlinePanel,
-    MultiFieldPanel,
-    PageChooserPanel,
-)
-from wagtail.api import APIField
-from wagtail.fields import RichTextField, StreamField
+from wagtail.admin.panels import FieldPanel
+
+# from wagtail.api import APIField
 from wagtail.models import Orderable
 
-from ..blocks import (
-    QuestionBlock,
-    SourceMediaBlock,
-)
-from ..serializers import (
-    KeyStageSerializer,
-    SourceSerializer,
-    ThemeSerializer,
-    TimePeriodSerializer,
-)
+# TODO serializers and API
 
 # TODO: sort panel order on promote tabs
 
@@ -189,14 +155,6 @@ class BaseKeyStageTag(Orderable):
         abstract = True
 
 
-class TeachingResourcePageKeyStageTag(BaseKeyStageTag):
-    page = ParentalKey(
-        "education.TeachingResourcePage",
-        on_delete=models.CASCADE,
-        related_name="education_keystage_tags",
-    )
-
-
 class BaseTimePeriodTag(Orderable):
     time_period = models.ForeignKey(
         "education.TimePeriod",
@@ -213,22 +171,6 @@ class BaseTimePeriodTag(Orderable):
         abstract = True
 
 
-class TeachingResourcePageTimePeriodTag(BaseTimePeriodTag):
-    page = ParentalKey(
-        "education.TeachingResourcePage",
-        on_delete=models.CASCADE,
-        related_name="education_time_period_tags",
-    )
-
-    time_period = models.ForeignKey(
-        "education.TimePeriod",
-        on_delete=models.CASCADE,
-        related_name="+",
-        verbose_name=_("time period"),
-        limit_choices_to={"available_for_resources": True},
-    )
-
-
 class BaseThemeTag(Orderable):
     theme = models.ForeignKey(
         "education.Theme",
@@ -243,886 +185,3 @@ class BaseThemeTag(Orderable):
 
     class Meta:
         abstract = True
-
-
-class TeachingResourcePageThemeTag(BaseThemeTag):
-    page = ParentalKey(
-        "education.TeachingResourcePage",
-        on_delete=models.CASCADE,
-        related_name="education_theme_tags",
-    )
-
-
-class EducationSessionPageKeyStageTag(BaseKeyStageTag):
-    page = ParentalKey(
-        "education.EducationSessionPage",
-        on_delete=models.CASCADE,
-        related_name="education_keystage_tags",
-    )
-
-
-class EducationSessionPageTimePeriodTag(BaseTimePeriodTag):
-    page = ParentalKey(
-        "education.EducationSessionPage",
-        on_delete=models.CASCADE,
-        related_name="education_time_period_tags",
-    )
-
-
-class EducationSessionPageThemeTag(BaseThemeTag):
-    page = ParentalKey(
-        "education.EducationSessionPage",
-        on_delete=models.CASCADE,
-        related_name="education_theme_tags",
-    )
-
-
-class Source(Orderable):
-    page = ParentalKey(
-        "education.TeachingResourcePage",
-        on_delete=models.CASCADE,
-        related_name="sources",
-    )
-
-    source_title = models.CharField(
-        verbose_name=_("source title"),
-        help_text=_("A unique, descriptive title for the source."),
-        blank=True,
-        max_length=160,
-    )
-
-    # Source Media — choose one type; caption is always included
-    source_media = StreamField(
-        [
-            (
-                "source_image",
-                wagtail.blocks.StructBlock(
-                    [
-                        (
-                            "image",
-                            APIImageChooserBlock(
-                                rendition_size="max-900x900",
-                                help_text=_("An image for the source."),
-                            ),
-                        ),
-                        (
-                            "caption",
-                            APIRichTextBlock(
-                                features=["bold", "italic"],
-                                required=False,
-                                help_text=_("Caption displayed below the image."),
-                            ),
-                        ),
-                    ],
-                    label=_("Image"),
-                ),
-            ),
-            (
-                "source_media",
-                SourceMediaBlock(
-                    help_text=_("A piece of audio or video media for the source."),
-                ),
-            ),
-            (
-                "source_youtube",
-                wagtail.blocks.StructBlock(
-                    [
-                        (
-                            "youtube",
-                            YouTubeBlock(
-                                help_text=_("A YouTube video for the source."),
-                            ),
-                        ),
-                        (
-                            "caption",
-                            APIRichTextBlock(
-                                features=["bold", "italic"],
-                                required=False,
-                                help_text=_("Caption displayed below the video."),
-                            ),
-                        ),
-                    ],
-                    label=_("YouTube"),
-                ),
-            ),
-        ],
-        verbose_name=_("source media"),
-        help_text=_(
-            "Choose one media type for this source. A caption can be added for each."
-        ),
-        blank=True,
-        null=True,
-    )
-
-    # Source link — choose internal or external
-    source_media_featured_link = StreamField(
-        [
-            (
-                "internal_page",
-                APIPageChooserBlock(
-                    label=_("Internal page"),
-                    page_type="wagtailcore.Page",
-                    help_text=_("Link to a page published on the site"),
-                ),
-            ),
-            (
-                "external_link",
-                wagtail.blocks.StructBlock(
-                    [
-                        (
-                            "url",
-                            wagtail.blocks.URLBlock(
-                                help_text=_(
-                                    "Link to a resource on a 3rd party platform (e.g. mapping tool)"
-                                ),
-                            ),
-                        ),
-                    ],
-                    label=_("External link"),
-                ),
-            ),
-        ],
-        verbose_name=_("source media featured link"),
-        help_text=_("Choose internal or external link for this source"),
-        blank=True,
-        max_num=1,
-    )
-
-    # Source description
-    source_description = StreamField(
-        [
-            (
-                "source_description",
-                APIRichTextBlock(features=settings.RESTRICTED_RICH_TEXT_FEATURES),
-            )
-        ],
-        verbose_name=_("source description"),
-        help_text=_(
-            "An optional free text field to add in a fuller description of the source."
-        ),
-        blank=True,
-        null=True,
-    )
-
-    # Source questions
-    source_question = StreamField(
-        QuestionBlock(
-            verbose_name=("source question"),
-            help_text=("A series of questions relating to each source."),
-        ),
-        blank=True,
-    )
-
-    panels = [
-        FieldPanel("source_title"),
-        FieldPanel("source_media"),
-        FieldPanel("source_media_featured_link"),
-        FieldPanel("source_description"),
-        FieldPanel("source_question"),
-    ]
-
-
-class CurriculumConnection(Orderable):
-    page = ParentalKey(
-        "education.TeachingResourcePage",
-        on_delete=models.CASCADE,
-        related_name="curriculum_connections",
-    )
-
-    key_stage = models.ForeignKey(
-        "education.KeyStage",
-        on_delete=models.CASCADE,
-        related_name="+",
-        verbose_name=_("Key stage"),
-        help_text=_("The key stage for this curriculum connection."),
-    )
-
-    description = RichTextField(
-        features=["bold", "italic", "link", "ul"],
-        verbose_name=_("curriculum connection description"),
-        help_text=_("Add the curriculum connection description."),
-        blank=True,
-        null=True,
-    )
-
-    panels = [
-        FieldPanel("key_stage"),
-        FieldPanel("description"),
-    ]
-
-
-class TeachingResourcePage(
-    PublishedDateMixin, RequiredHeroImageMixin, BasePageWithRequiredIntro
-):
-    """A page to display a teaching resource"""
-
-    parent_page_types = [
-        "education.TeachingResourcesListingPage",
-    ]
-
-    enquiry_question = models.CharField(
-        verbose_name=_("enquiry question"),
-        blank=True,
-        max_length=160,
-    )
-
-    @cached_property
-    def key_stages(self):
-        return [
-            tag.key_stage
-            for tag in self.education_keystage_tags.select_related("key_stage")
-        ]
-
-    @cached_property
-    def time_periods(self):
-        return [
-            tag.time_period
-            for tag in self.education_time_period_tags.select_related("time_period")
-        ]
-
-    @cached_property
-    def themes(self):
-        return [tag.theme for tag in self.education_theme_tags.select_related("theme")]
-
-    @cached_property
-    def key_stage(self):
-        return self.key_stages[0] if self.key_stages else None
-
-    @cached_property
-    def time_period(self):
-        return self.time_periods[0] if self.time_periods else None
-
-    @cached_property
-    def theme(self):
-        return self.themes[0] if self.themes else None
-
-    # Body
-    sources_title = models.CharField(
-        verbose_name=_("sources title"),
-        help_text=_(
-            "Title of the main section of the page. In most cases ‘Investigate the sources’"
-        ),
-        blank=True,
-        max_length=160,
-    )
-
-    sources_introduction = RichTextField(
-        features=["bold", "italic", "link"],
-        verbose_name=_("sources introduction"),
-        help_text=_("Optional text field to provide an introduction to the sources."),
-        blank=True,
-        null=True,
-    )
-
-    # Teacher’s Notes*
-    teachers_notes = RichTextField(
-        features=settings.RESTRICTED_RICH_TEXT_FEATURES,
-        verbose_name=_("teachers notes"),
-        help_text=_(
-            "A general overview of what the resource contains and how it can be used."
-        ),
-        blank=True,
-        null=True,
-    )
-
-    # Extension activities
-    extension_activities = StreamField(
-        [
-            (
-                "paragraph",
-                APIRichTextBlock(features=settings.RESTRICTED_RICH_TEXT_FEATURES),
-            ),
-            ("sub_heading", SubHeadingBlock()),
-            ("featured_page", FeaturedPageBlock()),
-            ("featured_external_link", FeaturedExternalLinkBlock()),
-        ],
-        verbose_name=_("extension activities"),
-        help_text=_(
-            "Optional section where editors can add extra activities for teachers to try with their pupils."
-        ),
-        blank=True,
-        null=True,
-    )
-
-    # Background information
-    background_information = StreamField(
-        [
-            (
-                "paragraph",
-                APIRichTextBlock(features=settings.RESTRICTED_RICH_TEXT_FEATURES),
-            ),
-            ("sub_heading", SubHeadingBlock()),
-        ],
-        verbose_name=_("background information"),
-        help_text=_("Section providing historical context to the teaching resource."),
-        blank=True,
-        null=True,
-    )
-
-    # Further information
-    further_information_title = models.CharField(
-        max_length=255,
-        verbose_name=_("further information title"),
-        help_text=_("Title of the further information section."),
-        blank=True,
-        null=True,
-    )
-
-    further_information = StreamField(
-        [
-            (
-                "paragraph",
-                APIRichTextBlock(features=settings.RESTRICTED_RICH_TEXT_FEATURES),
-            ),
-            ("sub_heading", SubHeadingBlock()),
-            ("featured_external_link", FeaturedExternalLinkBlock()),
-            ("featured_page", FeaturedPageBlock()),
-        ],
-        verbose_name=_("further information"),
-        help_text=_("Section providing links to other useful information."),
-        blank=True,
-        null=True,
-    )
-
-    content_panels = (
-        BasePageWithRequiredIntro.content_panels
-        + RequiredHeroImageMixin.content_panels
-        + [
-            FieldPanel("enquiry_question"),
-            MultiFieldPanel(
-                [
-                    FieldPanel("sources_title"),
-                    FieldPanel("sources_introduction"),
-                    InlinePanel(
-                        "sources",
-                        label=_("Source"),
-                        heading=_("Sources"),
-                        min_num=1,
-                    ),
-                ],
-                heading=_("Sources"),
-            ),
-            FieldPanel("teachers_notes"),
-            InlinePanel(
-                "curriculum_connections",
-                label=_("Curriculum connection"),
-                heading=_("Connections to the curriculum"),
-                min_num=1,
-            ),
-            FieldPanel("extension_activities"),
-            FieldPanel("background_information"),
-            MultiFieldPanel(
-                [
-                    FieldPanel("further_information_title"),
-                    FieldPanel("further_information"),
-                ],
-                heading=_("Further information"),
-            ),
-        ]
-    )
-
-    promote_panels = (
-        BasePageWithRequiredIntro.promote_panels
-        + PublishedDateMixin.promote_panels
-        + [
-            InlinePanel(
-                "education_keystage_tags",
-                label=_("Key stage"),
-                heading=_("Key stages"),
-            ),
-            InlinePanel(
-                "education_time_period_tags",
-                label=_("Time period"),
-                heading=_("Time periods"),
-            ),
-            InlinePanel(
-                "education_theme_tags",
-                label=_("Theme"),
-                heading=_("Themes"),
-            ),
-        ]
-    )
-
-    api_fields = BasePageWithRequiredIntro.api_fields + [
-        APIField("hero_image"),
-        APIField("enquiry_question"),
-        APIField("key_stage", serializer=KeyStageSerializer()),
-        APIField("time_period", serializer=TimePeriodSerializer()),
-        APIField("theme", serializer=ThemeSerializer()),
-        APIField("key_stages", serializer=KeyStageSerializer(many=True)),
-        APIField("time_periods", serializer=TimePeriodSerializer(many=True)),
-        APIField("themes", serializer=ThemeSerializer(many=True)),
-        APIField("sources_title"),
-        APIField("sources_introduction", serializer=RichTextSerializer()),
-        APIField("teachers_notes"),
-        APIField("sources", serializer=SourceSerializer(many=True)),
-        APIField("curriculum_connection_description", serializer=RichTextSerializer()),
-        APIField("extension_activities"),
-        APIField("background_information"),
-        APIField("further_information_title"),
-        APIField("further_information"),
-    ]
-
-
-# TODO: make this more generic? This is lifted from EducationReadMoreLink just as PoC and I don't like that
-class RelatedEducationSessions(Orderable):
-    """Links to take users to related educaiton sessions"""
-
-    page = ParentalKey(
-        "education.EducationSessionPage",
-        on_delete=models.CASCADE,
-        related_name="related_education_sessions",
-    )
-
-    selected_page = models.ForeignKey(
-        "wagtailcore.Page",
-        on_delete=models.CASCADE,
-        related_name="+",
-        verbose_name=_("selected page"),
-        help_text=_("Navigation to other Education sessions"),
-    )
-
-    panels = [
-        PageChooserPanel("selected_page"),
-    ]
-
-    class Meta:
-        verbose_name = _("related education session")
-        ordering = ["sort_order"]
-
-
-class EducationSessionPage(
-    PublishedDateMixin, RequiredHeroImageMixin, BasePageWithRequiredIntro
-):
-    """A page to display an education session"""
-
-    parent_page_types = [
-        "education.EducationSessionsListingPage",
-    ]
-
-    @cached_property
-    def key_stages(self):
-        return [
-            tag.key_stage
-            for tag in self.education_keystage_tags.select_related("key_stage")
-        ]
-
-    @cached_property
-    def time_periods(self):
-        return [
-            tag.time_period
-            for tag in self.education_time_period_tags.select_related("time_period")
-        ]
-
-    @cached_property
-    def themes(self):
-        return [tag.theme for tag in self.education_theme_tags.select_related("theme")]
-
-    @cached_property
-    def key_stage(self):
-        return self.key_stages[0] if self.key_stages else None
-
-    @cached_property
-    def time_period(self):
-        return self.time_periods[0] if self.time_periods else None
-
-    @cached_property
-    def theme(self):
-        return self.themes[0] if self.themes else None
-
-    # Page title* - default
-
-    # Hero image* - mixin
-
-    # Introductory text* -  mixin
-
-    # Rich text formatting (bold, italics, links)
-
-    # Start date  and End date
-    # hlep taxt placeholder, should be put at start of panel if so
-
-    session_start_date = models.DateField(
-        verbose_name=_("start date"),
-        null=True,
-        help_text=_(
-            "If neither start nor end date is added this will default to 'All Year'"
-        ),
-    )
-
-    session_end_date = models.DateField(
-        verbose_name=_("end date"),
-        null=True,
-        help_text=_(
-            "If neither start nor end date is added this will default to 'All Year'"
-        ),
-    )
-
-    # Optional field - Not all sessions have a specific date/dates
-
-    # TODO: Default to ‘All year' if no start and end date is added.
-    # [tick box? date optional?]
-
-    # or
-
-    # pick a start and end date from a date picker.
-
-    # startDate - Schema.org Property
-
-    # endDate - Schema.org Property
-
-    # TODO:
-    # Key stage
-
-    # Pull from metadata
-
-    # snippets? hard coded? models?
-
-    # Time period
-
-    # Pull from metadata
-
-    # Theme
-
-    # Pull from metadata
-
-    # --------Event key details bar
-
-    # Session price*
-
-    # Whether the session is free or paid (note, most are free)
-
-    # Default to ‘Free to UK schools’ if no text added
-
-    # Price field (optional)
-    # Price detail (plain text) (optional - not without price)
-
-    # price - Schema.org Property
-
-    session_price = None  # Optional float
-
-    # TOFO:Add validation
-    # or is this just a preferred currency format
-    session_price_detail = models.CharField(  # TODO- conditionally required,
-        verbose_name=_("session price detail"),
-        help_text=_("An explanation of the price."),
-        blank=True,
-        max_length=160,
-    )
-
-    # Location*
-
-    session_location = None
-    # write migration to seed choices? Maybe not, these could change a lot given recent talks?
-
-    # Ability to add one or more locations to the session
-
-    # Note, duration of sessions can vary depending on location, so duration info is added with each location info
-
-    # e.g.
-
-    # Online
-
-    # The National Archives
-
-    # Your school
-
-    # The Discovery Museum
-    # Blandford Square, Newcastle upon Tyne, NE1 4JA
-
-    # Museum of Richmond
-    # 2nd Floor, Old Town Hall, Whittaker Avenue, Richmond, TW9 1TP.
-
-    # Select from:
-    # - Online’
-    # - ‘At the National Archives'
-    # - At your school
-
-    # OR:
-
-    # Venue name [free text]
-    # Venue address [optional
-    # First line
-    # Second line
-    # postcode]
-
-    # Duration [free text}
-
-    session_location_duration = models.CharField(
-        verbose_name=_("session duration"),
-        help_text=_("A clear description of the session duration."),
-        blank=True,
-        max_length=160,
-    )
-
-    # Region
-    # [Regions:
-    # South East and London
-    # South West
-    # Midlands
-    # North East
-    # North West ]
-
-    session_regions = None
-
-    # PostalAddress - Schema.org Type (postalCode)
-    session_postal_address = None
-    # Recommended for*
-
-    # Key stage link
-
-    # Pull from Key Stage metadata
-    session_reccomended_for = None  # required
-
-    # Duration
-
-    # How long each session is. This will vary depending on location so will pull info from the location info above.
-
-    # Shows in front end in format of:
-
-    # At [venue name] – [duration]
-
-    # e.g. Online – 1 hour
-
-    # At The National Archives – 2 hours
-
-    # At The Discovery Museum – 1 hour
-
-    # At your school – 1 to 2 hours
-
-    # Location text:
-    # either ‘online’ 'At the National Archives, or Venue Name from location info
-
-    # Duration text: duration from location info (see above)
-
-    session_duration = None
-
-    # Booking link
-
-    # Link to book tickets. In most cases the same Eventbrite link, but some will have their own specific link
-
-    session_booking_link = models.URLField(
-        null=True,
-        blank=True,
-        help_text="Link to booking page",
-        verbose_name="Booking link",
-    )
-
-    # On front end, text will say ‘Book now’
-
-    # Add link
-
-    # ----------Event section
-
-    # Event body
-    # Field name
-
-    # Event description*
-
-    # Main body of the page where the session and what to expect from it are described
-
-    # - Paragraph text
-    session_description = RichTextField(
-        features=["bold", "italic", "link", "ul"],
-        verbose_name=_("session description"),
-        help_text=_("Add the session description."),
-        blank=True,
-        null=True,
-    )
-    # - Heading
-    session_title = models.CharField(
-        verbose_name=_("session title"),
-        help_text=_("A unique, descriptive title for the session."),
-        blank=True,
-        max_length=160,
-    )
-    # - Partner logo
-
-    partner_logo = StreamField(
-        [
-            (
-                "partner_logo",
-                PartnerLogoChooserBlock(
-                    rendition_size="max-900x900",
-                    verbose_name=_("partner logo"),
-                    help_text=_("An image for the partner logo."),
-                ),
-            )
-        ],
-        blank=True,
-    )
-    # - Quote
-    session_quote = StreamField(
-        [
-            (
-                "session_quote",
-                QuoteBlock(
-                    help_text=_("A quote with an attribution related to the session."),
-                ),
-            )
-        ],
-        blank=True,
-    )
-    # - Inset text TODO: clarify this for better naming but for now
-    inset_text = RichTextField(
-        features=["bold", "italic", "link", "ul"],
-        verbose_name=_("inset text"),
-        help_text=_(
-            "Inset text - TODO ask for better decsription of this? where it sits etc"
-        ),
-        blank=True,
-        null=True,
-    )
-
-    # Connections to the curriculum - not the same as in resources? but assuming they're repeatable?
-
-    # Area headed ‘Connection to the curriculum' (title hard coded) where specific connections to the curriculum can be listed. Note these is more specific connections than the themes (e.g. might specify exam board), so is a free text area.
-
-    # Rich text - Italic, bold, bulleted lists, numbered lists, links
-    session_curriculum_connection_description = StreamField(
-        [
-            (
-                "session_curriculum_connection_description",
-                APIRichTextBlock(features=settings.RESTRICTED_RICH_TEXT_FEATURES),
-            )
-        ],
-        verbose_name=_("source description"),
-        help_text=_(
-            "An optional free text field to add in a fuller description of the source."
-        ),
-        blank=True,
-        null=True,
-    )
-
-    # Event highlights
-    # Nice to have. Not for MVP - easy to add though
-
-    # Same as Event highlights component
-
-    session_highlights = StreamField(
-        [("session_highlights", ImageGalleryBlock())],
-        blank=True,
-        help_text=_("Optional image gallery to show what to expect from the session."),
-        max_num=1,
-    )
-
-    # Contact us - contact us component?
-
-    # Area where users can find out where to get support with their booking
-
-    # Will be the same in most cases but some will have different contact info.
-
-    # Make them fill it in every time. Copy the Contact us component
-
-    # Text
-
-    # email
-
-    # phone number
-
-    # Newsletter to be handled in the frontend
-
-    # More education sessions  - handled by frontend? use cached property to get them from taxonomies?
-    # then allow override using the selected pages from related_education_sessions via RelatedEducationSessions
-
-    # TODO: think about adding key details tab ala /services/ds-wagtail/app/whatson/models/details.py: for key details bar lol
-
-    #     edit_handler = TabbedInterface(
-    #         [
-    #             ObjectList(content_panels, heading="Content"),
-    #             ObjectList(key_details_panels, heading="Key details"),
-    #             ObjectList(promote_panels, heading="Promote"),
-    #             ObjectList(BasePageWithRequiredIntro.settings_panels, heading="Settings"),
-    #         ]
-    #     )
-
-    content_panels = (
-        BasePageWithRequiredIntro.content_panels
-        + RequiredHeroImageMixin.content_panels
-        + [
-            FieldPanel("session_start_date"),  # TODO: default to all year
-            FieldPanel("session_end_date"),
-            MultiFieldPanel(
-                [
-                    MultiFieldPanel(
-                        [
-                            # FieldPanel("session_price"), #make this multifield panel?
-                            MultiFieldPanel(
-                                [
-                                    FieldPanel("session_price_detail"),
-                                ],
-                                heading=_("Session price"),
-                            ),
-                            MultiFieldPanel(
-                                [  # TODO: sort this out - location and duration linked
-                                    # does this need to be multifield? probs not look at spec to double chec no other firleds
-                                    FieldPanel("session_location_duration"),
-                                ],
-                                heading=_("Session location"),
-                            ),
-                            # FieldPanel("session_reccomended_for"),
-                            # FieldPanel("session_duration"),
-                            FieldPanel("session_booking_link"),
-                        ],
-                        heading=_("Session description"),
-                    ),
-                    # FieldPanel("curriculum_connection_description"),
-                    # FieldPanel("session_highlights"),
-                ],
-                heading=_("Session key details bar"),
-            ),
-            MultiFieldPanel(
-                [
-                    MultiFieldPanel(
-                        [
-                            FieldPanel("session_description"),
-                            FieldPanel("session_title"),
-                            FieldPanel("partner_logo"),
-                            FieldPanel("session_quote"),
-                            FieldPanel("inset_text"),
-                        ],
-                        heading=_("Session description"),
-                    ),
-                    FieldPanel("session_curriculum_connection_description"),
-                    FieldPanel("session_highlights"),
-                ]
-            ),
-            InlinePanel(
-                "related_education_sessions",
-                heading=_("More education sessions"),
-                help_text=_(
-                    "Education sessions that are selected to be shown in the related education sesisons section"
-                ),
-            ),
-        ]
-    )
-
-    promote_panels = (
-        BasePageWithRequiredIntro.promote_panels
-        + PublishedDateMixin.promote_panels
-        + [
-            InlinePanel(
-                "education_keystage_tags",
-                label=_("Key stage"),
-                heading=_("Key stages"),
-            ),
-            InlinePanel(
-                "education_time_period_tags",
-                label=_("Time period"),
-                heading=_("Time periods"),
-            ),
-            InlinePanel(
-                "education_theme_tags",
-                label=_("Theme"),
-                heading=_("Themes"),
-            ),
-        ]
-    )
-
-    api_fields = BasePageWithRequiredIntro.api_fields + [
-        APIField("key_stage", serializer=KeyStageSerializer()),
-        APIField("time_period", serializer=TimePeriodSerializer()),
-        APIField("theme", serializer=ThemeSerializer()),
-        APIField("key_stages", serializer=KeyStageSerializer(many=True)),
-        APIField("time_periods", serializer=TimePeriodSerializer(many=True)),
-        APIField("themes", serializer=ThemeSerializer(many=True)),
-    ]
